Reorganized project structure: moved code to modular files
