+++ analysis/sp500.py
@@ -10,7 +10,7 @@ def sp500_performance():
     sp500 = yf.Ticker("^GSPC")
 
     end_date = datetime.now().date()
-    start_date = end_date - timedelta(days=2)
+    start_date = end_date - timedelta(days=3)
 
     hist = sp500.history(start=start_date, end=end_date)
