+++ config/chrome_options.py
@@ -3,7 +3,6 @@ from selenium import webdriver
 from selenium.webdriver.chrome.service import Service
 from selenium.webdriver.chrome.options import Options
 
-
 def chrome_option():
     """
     Configures and returns a Chrome WebDriver instance with a specific user profile.
@@ -22,7 +21,6 @@ def chrome_option():
     chrome_options.binary_location = os.getenv("CHROME_BINARY_PATH", "/Applications/Google Chrome.app/Contents/MacOS/Google Chrome")
     chrome_options.add_argument(f"user-data-dir={os.getenv('CHROME_USER_DATA_PATH', '/default/path/to/user-data')}")
     chrome_options.add_argument(f"profile-directory={os.getenv('CHROME_PROFILE', 'Default')}")
-
 
     # Define the ChromeDriver service (adjust path as necessary)
     service = Service(executable_path='/usr/local/bin/chromedriver')
