+++ wealthsimple/auth.py
@@ -1,0 +1,95 @@
+from dotenv import load_dotenv
+import os
+from selenium import webdriver
+from selenium.webdriver.chrome.service import Service
+from selenium.webdriver.common.by import By
+from selenium.webdriver.common.keys import Keys
+from selenium.webdriver.support.ui import WebDriverWait
+from selenium.webdriver.support import expected_conditions as EC
+from selenium.webdriver.chrome.options import Options
+from config.chrome_options import chrome_option
+load_dotenv()
+
+email = os.getenv("WEALTHSIMPLE_EMAIL")
+password = os.getenv("WEALTHSIMPLE_PASSWORD")
+
+driver = chrome_option()
+
+def login():
+    """
+    Logs in to Welthsimple
+    """
+    try:
+        # Navigate to WealthSimple
+        driver.get("https://my.wealthsimple.com/app/login?locale=en-ca")
+
+        WebDriverWait(driver, 30).until(
+            EC.presence_of_element_located((By.XPATH, "//input[@inputmode='email']"))
+        )
+
+        email_element = driver.find_element(By.XPATH, "//input[@inputmode='email']")
+
+        # Clears any pre-existing text in the box
+        email_element.clear()
+
+        email_element.send_keys(email)
+
+        WebDriverWait(driver, 10).until(
+        EC.presence_of_all_elements_located((By.CSS_SELECTOR, "input[type='password']"))
+        )
+
+        password_element = driver.find_element(By.CSS_SELECTOR, "input[type='password']")
+        
+        # Clears any pre-existing text in the box
+        password_element.clear()
+
+        # After entering in the password, it will click enter to sign in
+        password_element.send_keys(password + Keys.ENTER)
+
+        print("Sucessfully Logged in!")
+
+    except Exception as e:
+        print(f"Failed to log in: {e}")
+
+def navigate_to_home():
+    """
+    Navigates to the home page, where 
+    all the menus and accounts are located.
+    """
+    try:
+        # Navigate to home page where all the info is located, default page.
+        WebDriverWait(driver, 10).until(
+            EC.element_to_be_clickable((By.LINK_TEXT, 'Home'))
+        )
+
+        home_page = driver.find_element(By.LINK_TEXT, 'Home')
+
+        home_page.click()
+
+        print("Navigated to home page.")
+    except Exception as e:
+        print(f"Error navigating to home page: {e}.")
+def logout():
+    """
+    Logs out of Wealthsimple securely.
+    Args:
+        driver: Selenium WebDriver instance.
+    """
+    navigate_to_home()
+
+    try:
+        logout_menu = WebDriverWait(driver, 20).until(
+            EC.element_to_be_clickable((By.XPATH, '//button[@aria-label="Profile and settings" and @role="button"]'))
+        )
+
+        logout_menu.click()
+
+        logout_button = WebDriverWait(driver, 60).until(
+            EC.element_to_be_clickable((By.XPATH, "//p[text()='Log out']"))
+        )
+        logout_button.click()
+        print('Sucessfully logged out)')
+        return True
+    except Exception as e:
+        print(f'Error logging out: {e}')
+        return False
